# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lupa - Consumer Payment Authorization API (proposed)",
    "description": "A proposed API for the buyer's side of agentic payments: see the permissions that can take your money, set limits on them, give an AI agent bounded authority, approve what falls outside it. PayPal has no such API today; this prototype runs it on the PayPal sandbox. ◇ endpoints under /proposed/v1/me are the proposal. ● routes under /paypal call the sandbox. ◆ marks AI interpretation, ■ deterministic policy. Amounts are integer minor units. Full reference: https://github.com/apn201/Lupa/blob/main/docs/api-reference.md",
    "version": "0.1.0"
  },
  "paths": {
    "/proposed/v1/me/payment-permissions": {
      "get": {
        "tags": [
          "◇ proposed"
        ],
        "summary": "List payment permissions",
        "operationId": "list_permissions_proposed_v1_me_payment_permissions_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "attention",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Attention"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "◇ proposed"
        ],
        "summary": "Create an agent permission",
        "description": "Create an agent_authority permission. Delegate it with POST .../delegate.",
        "operationId": "create_agent_permission_proposed_v1_me_payment_permissions_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentPermissionIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proposed/v1/me/payment-permissions/{pid}": {
      "get": {
        "tags": [
          "◇ proposed"
        ],
        "summary": "Show payment permission details",
        "operationId": "get_permission_proposed_v1_me_payment_permissions__pid__get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proposed/v1/me/payment-permissions/{pid}/revoke": {
      "post": {
        "tags": [
          "◇ proposed"
        ],
        "summary": "Revoke a payment permission",
        "operationId": "revoke_proposed_v1_me_payment_permissions__pid__revoke_post",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "x-lupa-agent",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Lupa-Agent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proposed/v1/me/payment-permissions/{pid}/suspend": {
      "post": {
        "tags": [
          "◇ proposed"
        ],
        "summary": "Suspend a payment permission",
        "operationId": "suspend_proposed_v1_me_payment_permissions__pid__suspend_post",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "x-lupa-agent",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Lupa-Agent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proposed/v1/me/payment-permissions/{pid}/policy": {
      "put": {
        "tags": [
          "◇ proposed"
        ],
        "summary": "Set a permission's policy",
        "operationId": "put_policy_proposed_v1_me_payment_permissions__pid__policy_put",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proposed/v1/me/policies/default": {
      "get": {
        "tags": [
          "◇ proposed"
        ],
        "summary": "Show the default policy",
        "operationId": "get_default_proposed_v1_me_policies_default_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "◇ proposed"
        ],
        "summary": "Set the default policy",
        "operationId": "put_default_proposed_v1_me_policies_default_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proposed/v1/me/intent": {
      "post": {
        "tags": [
          "◇ proposed"
        ],
        "summary": "Draft a policy from a sentence",
        "description": "◆ Draft a policy from one sentence. Nothing is stored until PUT .../policy.",
        "operationId": "intent_proposed_v1_me_intent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proposed/v1/me/payment-permissions/{pid}/suggested-limit": {
      "get": {
        "tags": [
          "◇ proposed"
        ],
        "summary": "Get a suggested limit",
        "operationId": "suggested_limit_proposed_v1_me_payment_permissions__pid__suggested_limit_get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proposed/v1/me/payment-permissions/{pid}/explain": {
      "post": {
        "tags": [
          "◇ proposed"
        ],
        "summary": "Explain a permission",
        "operationId": "explain_proposed_v1_me_payment_permissions__pid__explain_post",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proposed/v1/me/payment-permissions/{pid}/delegate": {
      "post": {
        "tags": [
          "◇ proposed"
        ],
        "summary": "Delegate a permission to an agent",
        "operationId": "delegate_proposed_v1_me_payment_permissions__pid__delegate_post",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DelegateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proposed/v1/me/delegations/{did}": {
      "delete": {
        "tags": [
          "◇ proposed"
        ],
        "summary": "Revoke a delegation",
        "operationId": "revoke_delegation_proposed_v1_me_delegations__did__delete",
        "parameters": [
          {
            "name": "did",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Did"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proposed/v1/me/payment-permissions/{pid}/requests": {
      "post": {
        "tags": [
          "◇ proposed"
        ],
        "summary": "Create a payment request",
        "description": "A payment request arrives. ◆ assess, ■ decide, ● execute.",
        "operationId": "payment_request_proposed_v1_me_payment_permissions__pid__requests_post",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "x-lupa-agent",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Lupa-Agent"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentRequestIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proposed/v1/me/requests/{rid}/decision": {
      "post": {
        "tags": [
          "◇ proposed"
        ],
        "summary": "Submit an agent decision",
        "description": "An agent submits its own verdict. Stored as an AI verdict; the engine still decides.",
        "operationId": "submit_decision_proposed_v1_me_requests__rid__decision_post",
        "parameters": [
          {
            "name": "rid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rid"
            }
          },
          {
            "name": "x-lupa-agent",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Lupa-Agent"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Assessment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proposed/v1/me/requests/{rid}/resolve": {
      "post": {
        "tags": [
          "◇ proposed"
        ],
        "summary": "Resolve a held request",
        "operationId": "resolve_proposed_v1_me_requests__rid__resolve_post",
        "parameters": [
          {
            "name": "rid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proposed/v1/me/requests": {
      "get": {
        "tags": [
          "◇ proposed"
        ],
        "summary": "List payment requests",
        "operationId": "requests_proposed_v1_me_requests_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "held",
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proposed/v1/me/events": {
      "get": {
        "tags": [
          "◇ proposed"
        ],
        "summary": "List events",
        "operationId": "events_proposed_v1_me_events_get",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 300,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentPermissionIn": {
        "properties": {
          "agent_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Agent Id"
          },
          "label": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Label",
            "description": "what the agent buys, e.g. 'Shopping'"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "marketplace"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "label"
        ],
        "title": "AgentPermissionIn"
      },
...
      "Envelope": {
        "properties": {
          "marker": {
            "type": "string",
            "title": "Marker",
            "default": "◇"
          },
          "proposed": {
            "type": "boolean",
            "title": "Proposed",
            "default": true
          },
          "data": {
            "title": "Data"
          },
          "paypal": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paypal"
          }
        },
        "type": "object",
        "title": "Envelope"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IntentIn": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 300,
            "minLength": 3,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "IntentIn"
      },
...
      "PolicyIn": {
        "properties": {
          "max_amount": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Max Amount",
            "description": "minor units"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "EUR"
          },
          "recurring": {
            "type": "string",
            "enum": [
              "allow",
              "require_approval",
              "block"
            ],
            "title": "Recurring",
            "default": "allow"
          },
          "amount_increase": {
            "type": "string",
            "enum": [
              "allow",
              "require_approval",
              "block"
            ],
            "title": "Amount Increase",
            "default": "require_approval"
          },
          "increase_tolerance_pct": {
            "type": "integer",
            "title": "Increase Tolerance Pct",
            "default": 20
          },
          "new_merchant": {
            "type": "string",
            "enum": [
              "allow",
              "require_approval",
              "block"
            ],
            "title": "New Merchant",
            "default": "require_approval"
          },
          "max_per_month": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Per Month"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "created_by": {
            "type": "string",
            "enum": [
              "human",
              "ai_draft_accepted"
            ],
            "title": "Created By",
            "default": "human"
          },
          "intent_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent Text"
          }
        },
        "type": "object",
        "required": [
          "max_amount"
        ],
        "title": "PolicyIn"
      },
      "ResolveIn": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "approve",
              "reject"
            ],
            "title": "Action"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "ResolveIn"
      },
...
    }
  }
}